with boundary probelms

+++ RangeScanOO/Edge.h
@@ -6,8 +6,8 @@ class Edge
 {
 public:
 	float x_;					// 边的上端点x 的屏幕空间坐标（左上角为原点）
-	int y_up;					// 上端点的y
-	int y_down;					// 下端点的y
+	float y_up;					// 上端点的y
+	float y_down;					// 下端点的y
 	float z_up_;				// 上端点的z
 	int poly_id_;				// 所属多边形id
 	int edge_id_;
